Update demo scene for the new Entity model
